[SCRUM-66] rename: auth 패키지 내 구조 정리
