+++ src/main/java/com/projectlyrics/server/domain/auth/entity/enumerate/AuthProvider.java
@@ -1,0 +1,7 @@
+package com.projectlyrics.server.domain.auth.entity.enumerate;
+
+public enum AuthProvider {
+    KAKAO,
+    APPLE,
+    ;
+}
