--- src/main/java/com/projectlyrics/server/domain/auth/external/AuthProvider.java
@@ -1,7 +1,0 @@
-package com.projectlyrics.server.domain.auth.external;
-
-public enum AuthProvider {
-    KAKAO,
-    APPLE,
-    ;
-}
