--- src/main/java/com/projectlyrics/server/domain/auth/external/kakao/dto/KakaoAccount.java
@@ -1,6 +1,0 @@
-package com.projectlyrics.server.domain.auth.external.kakao.dto;
-
-public record KakaoAccount(
-        String email
-) {
-}
