+++ src/main/java/com/projectlyrics/server/domain/auth/service/kakao/dto/KakaoAccount.java
@@ -1,0 +1,6 @@
+package com.projectlyrics.server.domain.auth.service.kakao.dto;
+
+public record KakaoAccount(
+        String email
+) {
+}
