+++ src/main/java/com/projectlyrics/server/domain/user/entity/User.java
@@ -2,7 +2,7 @@ package com.projectlyrics.server.domain.user.entity;
 
 import com.projectlyrics.server.domain.auth.entity.Auth;
 import com.projectlyrics.server.domain.common.entity.EntityCommonField;
-import com.projectlyrics.server.domain.auth.external.AuthProvider;
+import com.projectlyrics.server.domain.auth.entity.enumerate.AuthProvider;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
