+++ src/main/java/com/projectlyrics/server/domain/user/usecase/command/UserAuthUseCase.java
@@ -1,7 +1,7 @@
 package com.projectlyrics.server.domain.user.usecase.command;
 
 import com.projectlyrics.server.domain.user.dto.response.LoginResponse;
-import com.projectlyrics.server.domain.auth.external.dto.request.UserLoginRequest;
+import com.projectlyrics.server.domain.auth.dto.request.UserLoginRequest;
 
 public interface UserAuthUseCase {
