+++ src/main/java/com/projectlyrics/server/global/configuration/FeignClientConfig.java
@@ -1,7 +1,7 @@
 package com.projectlyrics.server.global.configuration;
 
-import com.projectlyrics.server.domain.auth.external.kakao.KakaoApiClient;
-import com.projectlyrics.server.domain.auth.external.kakao.KakaoAuthApiClient;
+import com.projectlyrics.server.domain.auth.service.kakao.KakaoApiClient;
+import com.projectlyrics.server.domain.auth.service.kakao.KakaoAuthApiClient;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
